fix(xdata): Fix injections for XDATA blocks
The XDATA construct defaults to XML for content, but this can be
overriden using the MimeType keyword.  We used to match the various
mime-types for things like Markdown or JSON and explicitly set the
@injection.language and then had a "fallback" rule that would set it to
XML.  However, in Zed, this would result in two layers being created,
one for the matches langauge and one for XML -- this then caused
various rendering oddities.

In order to make these distinct, we had to follow the pattern we
established for methods that specify external languages with the
Language keyword.  Now, for (default) XML blocks, we use a
xdata_body_content_xml node and for other content, we use
xdata_body_content_any.  This makes it possible to create the necessary
queries in injections.scm that can properly match only the correct
language.