fix(core grammar): fix set and kill statements to be more permissive
Since oref_method is indistinguishable from a multidimensional property,
we must not try to force a property in set target, becuase that makes it
so valid set targets such as myarray.%Get(index).md(subscript) are marked
as invalid, since oref_method has precedence. Instead, we will allow for this,
and this must be checked semantically. Also, added kill statement tests.