fix(scanner,grammar): fix xecute statements and scanner to handle byref
Fixed xecute statements to accept valid xecute arguments that were previously
marked as invalid, such as xecute x,y. Additionally, adjusted the scanner
to account for byref values. In the case where xecute arguments are not
wrapped in parens, byref values are not allowed. However, if there was
whitespace, such as xecute x, .y, the scanner was previously eating the
., causing this to be marked as valid. To fix this, I added a new token
that stops the lexer  from advancing forward and eating the . in the case
that we are inside an xecute argument. This strategy was necessary, because
we can't stop the lexer from moving forward in all scenarios, as that
would cause valid dotted statements to give us errors.

Additionally, I added tests for xecute statements.