fix(scanner,grammar): fix embedded sql, html, and js For embedded sql, the scanner was allowing a zero width token, which is unnecessary. I adjusted the logic so the EMBEDDED_SQL_MARKER token is only returned if there is actually a marker. This removes the zero width  token problem.
For embedded html, I added the logic to include a marker and to verify
that the end part is in reversed order.

Although docs say embedded js is the exact same as html, there is one
special case where that is not true: when there is a '>' in your js.
To fix this, you  have to add a marker (explained on community posts).
I added logic that changes the way the scanner parses embedded js if there
is a marker. This makes it so it can parse js expressions that have '>' in them.

Additionally, I added unit tests for all of these.