fix(grammar,scanner):fix pound_def statements In the case where a #define was not followed by a newline, the termination rather than just newline.
Additionally, the rule that a ##continue and ##; comment cannot be
on the same line was not enforced, so I changed the scanner to enforce this.

Lastly, since I was using the scanner to check for termination, I had to use the
scanner to check that with the ##continue, there was exactly one new_line, and the
scanner would consume this so that new line wouldn't be registered as a termination.
I added unit tests for both valid and invalid #define statements.